database

+++ WebBanTra/WebBanTra/Models/ChiTietDNH.cs
@@ -17,7 +17,7 @@ namespace WebBanTra.Models
         public int MaDNH { get; set; }
         public int MaSP { get; set; }
         public Nullable<int> SoLuongNhap { get; set; }
-        public Nullable<decimal> GiaNhap { get; set; }
+        public Nullable<double> GiaNhap { get; set; }
     
         public virtual DonNhapHang DonNhapHang { get; set; }
         public virtual SanPham SanPham { get; set; }
