+++ WebBanTra/WebBanTra/OOP/CartDetail.cs
@@ -8,7 +8,7 @@ namespace WebBanTra.OOP
     {
         public int MaSP { get; set; }
         public string TenSP { get; set; }
-        public decimal Gia { get; set; }
+        public double Gia { get; set; }
         public int SoLuong { get; set; }
         public int MaDM { get; set; }
         public string LinkAnh { get; set; }
